GG-18787: [ML] Hyperparameter auto tuning, Phase 1 (#85)
* added support of parallel CV and HP tuning
* implemented different learning strategies
* preprocessing tuning
* adaptation Pipeline API for CV and HP

+++ modules/ml/src/main/java/org/apache/ignite/ml/preprocessing/normalization/NormalizationTrainer.java
@@ -53,9 +53,9 @@ public class NormalizationTrainer<K, V> implements PreprocessingTrainer<K, V> {
      * @param p The given value.
      * @return The Normalization trainer.
      */
-    public NormalizationTrainer<K, V> withP(int p) {
+    public NormalizationTrainer<K, V> withP(double p) {
         assert p > 0;
-        this.p = p;
+        this.p = (int)p;
         return this;
     }
 }
