+++ modules/ml/src/main/java/org/apache/ignite/ml/regressions/logistic/LogisticRegressionSGDTrainer.java
@@ -157,8 +157,8 @@ public class LogisticRegressionSGDTrainer extends SingleLabelDatasetTrainer<Logi
      * @param maxIterations The parameter value.
      * @return Model with new max number of iterations before convergence parameter value.
      */
-    public LogisticRegressionSGDTrainer withMaxIterations(int maxIterations) {
-        this.maxIterations = maxIterations;
+    public LogisticRegressionSGDTrainer withMaxIterations(double maxIterations) {
+        this.maxIterations = (int)maxIterations;
         return this;
     }
 
@@ -168,8 +168,8 @@ public class LogisticRegressionSGDTrainer extends SingleLabelDatasetTrainer<Logi
      * @param batchSize The size of learning batch.
      * @return Trainer with new batch size parameter value.
      */
-    public LogisticRegressionSGDTrainer withBatchSize(int batchSize) {
-        this.batchSize = batchSize;
+    public LogisticRegressionSGDTrainer withBatchSize(double batchSize) {
+        this.batchSize = (int)batchSize;
         return this;
     }
 
@@ -179,8 +179,8 @@ public class LogisticRegressionSGDTrainer extends SingleLabelDatasetTrainer<Logi
      * @param amountOfLocIterations The parameter value.
      * @return Trainer with new locIterations parameter value.
      */
-    public LogisticRegressionSGDTrainer withLocIterations(int amountOfLocIterations) {
-        this.locIterations = amountOfLocIterations;
+    public LogisticRegressionSGDTrainer withLocIterations(double amountOfLocIterations) {
+        this.locIterations = (int)amountOfLocIterations;
         return this;
     }
